fix: simplify directory structure
For the link checker to work, the directory structure must mirror the
directory path structure of the final site.

Since we had two parallel directory structures, `source`,
containing erb, and `source/documentation`, containing markdown,
and those structures were slightly inconsistent, we ended up in the case
where a relative link in the markdown could not be resolved
by the link checker, even though the link worked when the documentation
was published.

There is really no need to have a parallel file structure just to
include markdown files. Middleman (which the tech docs template is based
on) will happily mix erb files with the
extension `.html.erb` or `.html.md.erb` and markdown files with the
extension `.html.md`. In fact, we don't rely on erb at all, so we can
just name our files `.html.md`.